TEMP FIX: Rabbit slime transition to dodge
There's an issue caused by SteeringAgents reached goal flag making the
velocity to zero which in turn causes isMoving to be set to zero.

This forces a transition back to idle. Problem is this takes precedence
over a transition to dodge. I've disabled the transition back to idle
which should look fine for now as the Rabbit Slime is rarely ever
idling.